clean up of the config
